requirements.txt
python-dotenv==0.6.4
dropbox==7.2.1
click==6.7